+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - examples dataset and code-snippets
 - improved function to load datasets at LOSTDataset
 - new function `segmentation_to_lost` to convert pixelmaps to lost-annotations
+- new function `crop_components` to crop dataset based on annotations
 ## [0.0.0] - 2021-10-26
 ### Added
 - First version